% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle}
\alias{cycle}
\title{Data for pruning cycle graph: Mango tree steps (through 3 cycles)}
\format{
A data frame of 15 x 4

\describe{
\item{Debut}{Date of the beginning of the step}
\item{Fin}{Date of the end of the step}
\item{Etape}{Name/id of the step}
\item{Cycle}{Cycle id}
}
}
\usage{
cycle
}
\description{
Data for pruning cycle graph: Mango tree steps (through 3 cycles)
}
\keyword{datasets}
