% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_labels}
\alias{date_labels}
\title{Data for pruning cycle graph: Date labels for the x axis}
\format{
A data frame of 47 x 3

\describe{
\item{pas}{date break}
\item{annee}{turn the year into a general identifier}
\item{etiquette}{date label, with a month format + the year id once per year}
}
}
\usage{
date_labels
}
\description{
Data for pruning cycle graph: Date labels for the x axis
}
\keyword{datasets}
