% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_taille}
\alias{date_taille}
\title{Data for pruning cycle graph: Pruning dates}
\format{
A data frame of 4 x 6

\describe{
\item{Taille}{Type of pruning}
\item{Date_taille}{Pruning date}
\item{Depart}{Beginning of the arrow}
\item{Pointe}{End of the arrow}
\item{pos_img}{Position of the shears image}
\item{img}{Path to the shears image}
}
}
\usage{
date_taille
}
\description{
Data for pruning cycle graph: Pruning dates
}
\keyword{datasets}
