% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run APPNAME application}
\usage{
run_app(lang = "fr")
}
\arguments{
\item{lang}{interface langage : english (en) or french (fr)}
}
\description{
Warning: beware of launching the app in a web browser and not in the RStudio viewer
}
