% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taille}
\alias{taille}
\title{Pruning essay, from 2007 to 2021}
\format{
A data frame of 6 063 x 8

\describe{
\item{arbre}{tree identifier}
\item{bloc}{block of trees}
\item{Annee}{Year of the observation}
\item{Taille}{treatment: winter pruning, summer pruning or no pruning}
\item{X, Y}{Tree coordinates}
\item{Mesure, Valeur}{Either the production in kilograms (masse) or the number of fruits (nbfruit) or the mean fruit weight in grams (masse_fruit) per tree per year of harvest}
}
}
\usage{
taille
}
\description{
Pruning essay, from 2007 to 2021
}
\keyword{datasets}
