% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{textesUI}
\alias{textesUI}
\title{Translation texts for the UI}
\format{
3 columns for identifier, french and english
}
\usage{
textesUI
}
\description{
Translation texts for the UI
}
\keyword{datasets}
