% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variete}
\alias{variete}
\title{Varietal essay}
\format{
data.frame 2,814 x 7
\describe{
\item{arbre}{tree identifier}
\item{cultivar}{mango tree cultivar}
\item{Annee}{Year of the observation}
\item{X, Y}{Mtree coordinates}
\item{Mesure, Valeurs}{Either the production in grams (masse) or the number of fruits (nbfruit) or the mean fruit weight in grams (masse_fruit) per tree per year of harvest}
}
\usage{
variete
}
\description{
Varietal essay
}
\keyword{datasets}
