% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variete}
\alias{variete}
\title{Varietal essay, from 2010 to 2022}
\format{
A data frame of 4 824 x 7

\describe{
\item{arbre}{tree identifier}
\item{cultivar}{mango tree cultivar}
\item{Annee}{Year of the observation}
\item{X, Y}{Tree coordinates}
\item{Mesure, Valeur}{Either the production in kilograms (masse) or the number of fruits (nbfruit) or the mean fruit weight in grams (masse_fruit) per tree per year of harvest}
}
}
\usage{
variete
}
\description{
Varietal essay, from 2010 to 2022
}
\keyword{datasets}
